+++ changelog.md
@@ -21,6 +21,7 @@
   - implemented API to retrieve historical market data
   - do not query specification fields until it is received in TerminalState
   - upgraded CopyFactory API to 2.1.1
+  - increased default demo account request timeout to 240 seconds
 
 12.3.1
   - added feature to unsubscribe from market data (remove symbol from market watch)